Disabled tmux auto-connect/disconnect
Detaching from tmux was correctly exiting the SSH connection. However,
probably due to the use of `exec`, the tmux server was not maintained
afterwards.

Furthermore, it wasn't making so much sense to systematically enter
a tmux session. A future change, better, would be to open a panel
to select the tmux session wanted on connection.

Detaching the tmux session will thus not disconnect use anymore.
However, we could use the SSH binding to exit the connection
(`<Return>~.`), and maybe remap it, to directly exit the connection
without even detaching from tmux.